Do not lookup by name when unloading but by DeclId_t.
Fixes spurrious, occasional (platform-specific) failures in
roottest/cling/reload/reloadvector:
Trying to unloading the global function reloadvector(), the callback told TCling
that an unload was about to happen.
TCling called FindObject("reloadvector") which (on some platforms) created new
elements and (sometimes) deserialized decls, which in turn appended a new
transaction which then failed the assert that the transaction to be unloaded is
indeed the last one - because a new one was appended in the meantime.