GUI: Introduce a slightly new look and feel
Introduce a more "modern" look and feel by flattening the frames and widgets (avoiding double border) and using a more grey-ish color scheme.
To select the new style, add the "flat" keyword in the Gui.Style entry of the $(ROOTSYS)/etc/system.rootrc (or in the $(HOME)/.rootrc) file. For example:
Gui.Style:                  modern-flat

N.B. The default style is kept as "modern" for the time being