Separate the concept of enabling parsing on demand from when the header should be parsed.
TInterpreter::SetClassAutoParsing still completely disable the auto parsing
mechanishm, when set to true, the payload is parsed *immediately* upon library
load.

   TInterpreter::SuspendAutoParsing autoParseRaii(gInterpreter);

request the interpreter does *not* consider the header files that have been
registered to be parsed on demand but have not been (auto)parsed yet. This
allows to delay even further the parsing of those headers.