Fix ROOT-6796: propagate attrs to classes also if selctd through typedefs
the bug reported a class selected via its typedef and unwanted in the
rootmap.
The solution to this bug also represents a performance optimisation.
Previously, when it was time to annotate decls with the attributes
coming from their selection, the selection rule was found back matching
again against the class name. Now the rule scanner keeps a map of
pointers to canonical decls and BaseSelectionRules for an immediate
match.