Re-organise fwd decls in rootmaps: faster/less memory needed
the procedure leading to autogenerated code for forward declarations in
rootmaps has been improved.
The code now does not contain any duplicate under any condition within
the same rootmap. Furthermore, the preprocessing step which happens
in TCling can eliminate more duplications across different rootmaps.
This results in a faster processing of the rootmaps because cling has
to interpret less code. This is important for example at startup,
especially in presence of large sets of rich rootmaps (e.g. CMS, Atlas).
Since cling interprets less code, less memory is used. The overall
amount of memory saved considering the set of rootmaps of CMS is
of the order of about 5 MB.