Reduce mem usage of getting norm name in TCling::UpdateClassInfoWithDecl
The method
void TCling::UpdateClassInfoWithDecl(const void* vTD)
updates TClass instances with class infos if needed, i.e. when the
classes are in the kNoInfo, kForwardDeclared or kEmulated state.
In order to do this it, before this commit, it built the normalised name
and then used it as a key to find the TClass instance.
The normalisation is a procedure which consumes memory and it accounted
to 20-40 MB RSS, depending on the workflow, for standard CMS jobs (the
effect was comparable on systems characterised by a comparable complexity,
such as Atlas).
The idea is therefore to calculate the normalised name only if there is a
chance to find the TClass instance in the ROOT typesystem.
This is formalised exploiting the fact that one part of the names in ROOT
and in clang is the same, no matter what, and this is the "decl name".
For example, for a template A::B::C<...> the decl name would be C.
We now collect the "DeclNames" for those TClassInstances which are
created in kNoInfo, kForwardDeclared and kEmulated states in a catalogue.
If the declname of the decl is not in this catalogue, we do not bother
getting its normalised name: we know we would not find it anyhow.
Relevant details of the implementation:
o At the heart of the catalogue we have a simple unordered_set<string>
o The access to this catalogue is thread safe: concurrent additions and
reads simultaneous to additions are possible. The guard is a simple spin
lock guard relying on an atomic flag.