No a.parse with a template name coming from key which is atemplate instance
If an autoparse is requested for something which is a template instance name,
say A::B::MyClass<int>, the system is forced to autoparse for the template name,
A::B::MyClass, which may well be already in the system.
Now we check using cling if the template is not a fwd decl and is present in
the ast (which means also the pch).