XRootD-Support: workaround for the problem with 'private' headers
During a discussion with the XRootD main developers we agreed on the steps to gradually
remove the dependency from headers distributed in the 'private' RPM - not installed, for
example, on lxplus.
In the meantime, the workaround consists in importing under 'proof/xrdinc' the headers
incriminated, so that builds can be done also when include/xrootd/private is not installed.
The relevant legacy libraies are always installed, for the time being; the time scale for
their removal (probably end of this year) determines the time scale for the removal of
private header dependency.
This patch creates 'proof/xrdinc' with the relavant legacy headers and makes the relevant
changes in the configuration and (c)make files to deal with it.