Made TListOfFunctions thread safe
TlistOfFunctions can update itself over time, even from calls to
const functions. Since the class instances are available globally
they need to be threads safe. This required that all 'const' functions
take the global lock. In addition, a new iterator was needed which
takes the global lock on the call to Next().