Proof-PerfStat: adapt to dynamic startup
Add possibility to force setting the number of workers. This is useful when only a maximum is known,
as in  the case of dynamic startup. The TParameter<Int_t> is called 'PROOF_DynamicStartup' and the value
is the maximum number of workers (wmax). It is assumed that the ordinals are 0.0 ... 0.'wmax-1' .