Proof-Lite: add simulator for dynamic worker startup
The patch makes TProof::PollForNewWorkers virtual and adds an implementation of the same in TProofLite,
so that we can simulate dynamic startup, for packetizer testign purposes. The all thing is triggered
by the ROOT-rc

              'Proof.SimulateDynamicStartup'

which accepts a string value in the form "increase-step:max-num-workers", where 'increase-step' is the number
of new workers created at each call, and 'max-num-workers' is the maximum number of workers to start.
Once this is set, if a PROOF-Lite session with less workers than 'max-num-workers', during processing the
numbers of workers will be gradually increased to reach 'max-num-workers'.

The patch also re-organize the TProof code for dynamic startup to minimize duplications. In particular, the
part setting up the new workers environment has been separated into a new method SetupWorkersEnv, so that
it can used by TProofLite too.
Several bugs have been fixed in this code as a results of tests with examples requiring packages and loaded
macros, which were not used in the simple example used to develop this code.