Revert "Fix usage of thread_local on OSx 10.9 using the ROOT macro TTHREAD_TLS"
The macros in the ThreadLocalStorage.h are not the panacea.
For example, the standard C++ keyword "thread_local" is not equivalent to the
TTHREAD_TLS(xyz) syntax. Indeed, the presence of a static storage, it leads
to static static __thread constructs not digested on many platforms.

They work on kubuntu14 and SLC6, but they fail on osx10.9 and other platforms.
They should be used differently.

This reverts commit 5758705f205c197d0929ad0e4aeaf798f09f7fd8.