Fix python failure due to __1 being present (but not enough)
The presence of __1 is due to the import of a very small part of libcpp without
changing the code enough to avoid this __1.

The cause of the incomplete presence of __1 was a (quiet but complete) failure
in the setup/import part of ROOT.py leading to a dysfunctional PyROOT.

This was harder to track down as expected as the error which should have been as
described below is somehow suppressed when the code is executed as part of
the python import mechanism.  Which in this case, after parsing ROOT.py and
because of the statement
   sys.modules[ __name__ ] = ModuleFacade( sys.modules[ __name__ ] )
which customize the import mechanism for this modules, end up calling
   ModuleFacade.__getattr1 which the parameter __all__
which calls
   ModuleFacade.__finalSetup
which calls the problematic code (which is problematic because of the __1 existing but
cout not being in there):
   self.__dict__[ 'cout' ] = getattr( cppyy.gbl.std, '__1' ).cout

Note that the error suppressed are solely the one that happens within
the PyROOT C++ code underneath getattr( cppyy.gbl.std, '__1' ).cout.

This code ends properly executing
   (anonymous namespace)::LookupCppEntity (pyname=0x0, args=0x7fffe5fc5090) at /local2/pcanal/cint_working/rootcling/root/bindings/pyroot/src/RootModule.cxx:187

and in particular the line:

218           PyErr_Format( PyExc_AttributeError, "%s", name.c_str() );

with name containing ‘cout’ but nonetheless two things happens

  a) the message is not printed
  b) the function immediately exit (quietly) and never finishes the setup.

consequently PyROOT does not work in those circumstances and no error
message appears.

(Full disclosure this was stumbled upon debugging):

   cd roottest/cling/functionTemplate
   python testcint.py - -b

which rather than issuing the ‘right’ error message:

Traceback (most recent call last):
  File "testcint.py", line 1, in <module>
    from ROOT import *
  File "/home/pcanal/cint_working/rootcling/root/altc++11/lib/ROOT.py", line 416, in __getattr1
    self.__finalSetup()
  File "/home/pcanal/cint_working/rootcling/root/altc++11/lib/ROOT.py", line 521, in __finalSetup
    self.__dict__[ 'cout' ]
AttributeError: 'cout'

failed (because of the incomplete PyROOT setup) with:

Traceback (most recent call last):
  File "testcint.py", line 10, in <module>
    sortedMethods = [ item for item in t.__dict__.keys() if item[0:2] != '__' ]
NameError: name 't' is not defined