ROOT-7015 including some related changes:
Fix the reported crash by ensuring AddBranchToCache, DropBranchFromCache, SetCacheEntryRange,
StopCacheLearningPhase (from the TTree base) apply to the current tree of a TChain, not the
TChain itself. Also create a cache if necessary if one has not already been created; this avoids
the situation where the user could ask for a branch to be added to a TTree's cache and have the
call silently have no effect.

Changing the size of an existing TTreeCache, by called SetCacheSize with a different, non zero
size, is changed to avoid entirely recreating the cache and thereby restarting the learning phase.
It may still trigger the current cluster to be recached if the allowed size increases since more
baskets may fit in the cache.

Changed the return type of AddBranch, AddBranchToCache, SetCacheSize, SetCacheSizeAux, DropBranch,
DropBranchFromCache, SetCacheEntryRange, StopCacheLearningPhase in TFileCacheRead, TChain, TTree,
TTreeCache, TTreeCacheUnzip from void to Int_t to allow these methods to return an error indicator.
A number of Error() and one Warning() messages were also added to these methods.