Proof: single out merge and transfer-to-client time
This patch adds a measurement of the merge time, defined on the master from the moment the first
worker is done to the end of the query, and of the transfer time, the same quantity measured by
the client. The duration of the prepare phase, the transfer of the query definition from the client
to the master, and of the time taken by Terminate, are also recorded.
The patch also modifies TQueryResult so that the new time measurement are available for performance
studies. TQueryResult now gives something like this:

root [] p->GetQueryResults()->Print("F")
Collection name='TList', class='TList', size=1
+++
+++ #:1 ref:"session-pcphsft64-1427241610-787:q1" sel:ProofSimple finalized
+++        started:   Wed Mar 25 01:00:26 2015
+++        init:      0.202 sec
+++        process:   9.240 sec (CPU time: 198.0 sec)
+++        merge:     1.745 sec (5 mergers)
+++        transfer:  0.820 sec
+++        processed: 1000000 events (size: 0.000 MBs)
+++        rate:      108225.1 evts/sec
+++        # workers: 24
+++        results:   <PROOF_SandBox>/queries/session-pcphsft64-1427241610-787/1/query-result.root
+++        outlist:   1010 objects

The prepare and terminate phases are only show if longer than a millisec.

The patch also fixes some small issues about the exactness of some notification messages.