Update ROOT version files to v6.02/07 to follow new/old versioning.
Now, odd patch number are between tags; tags will always have even
patch numbers. Thus the next tag will be v6.02/08.

This allows users to determine whether they are building against a
tag or the ROOT sources after a tag, and it allows us to determine
whether a report is from a tagged build or a snapshot.