Properly support TStreamerInfo written by ROOT v4.00.
In (at least some) files written by v4.00, the TStreamerElement's type was not correct (or not
correctly recorded) such that the value of TStreamerElement::GetType was more accurate that
the type of the TStreamerElement object.  This conflicts with the optimization we recently made
that assumes that TStreamerBasicType points to a ... basic type.  Fix the problem by detecting
we are in the old case, and recreating the TStreamerElement with the 'correct' class (given
it's TStreamerElement::GetType).