ProofBench: add possibility to save derived objects created/drawn by TProofPerfAnalysis
TProofPerfAnalysis internally creates temporary histograms or graphs for basic visualisation.
This patch adds the possibility to save these objects to a file for further usage, for example
to create more elaborate plots.

Saving is enabled by a call to SetSaveResult, which can be used to change file and/or to disable:

    root [] TProofPerfAnalysis paf("perf_analysis_output.root")
    root [] paf.SetSaveResult()
    Drawn objects will be saved in file 'results.root'
    root [] paf.RatePlot()
    root [] paf.SetSaveResult("results-global.root", "UPDATE")
    Drawn objects will be saved in file 'results-global.root'
    root [] paf.RatePlot()
    root [] paf.SetSaveResult(0)
    Drawn objects saving disabled