Impose proper behaviour for unordered associative containers
this test was *erroneously* sometimes assuming ordering for unordered
containers.
This "oxymoron" did not have any effect on linux but it had some on mac:
the data read from disk differed from the reference. This was not because
of some corruption, but simply because different elements were compared.
Now order is imposed also for randomisation and histogram filling therewith
preventing such nasty effects.