Primer improvements including python filter for file inclusion
o Several improvements of the primer up to chapter 6 excluded
o Add rootDocPreprocessor.py script which is meant to be used to
"preprocess" the documentation. It applies a series of filters
to the input file, for example for including external files like
macros or code in general. The script was integrated in the
Makefile configuration and the inclusion of the macros is now
managed via directives in the text, namely @ROOT_INCLUDE_FILE
filename.