Use TClass::New in TSelector::GetSelector.
Since the selector inherits from TSelector (and hence TObject), it must have
a ClassDef and a dictionary (when compiled), so we can rely on it having a TClass.
Even if the selector is 'interpreted' a TClass will be created on the fly and
have a functional 'New' function (which goes through the ClassInfo).

No longer fall back on TSelectorCint as with Cling the 'interpreted' selector
is a valid C++ class class with a proper virtual table and can be used
exactly the same as a compiled selector.