Add THashTable::AddBefore to fix the ordering of the keys in a TDirectoryFile
Using THashList::AddBefore you could order the object in the list but not
within the buckets of the hash table resulting in a different ordering
depending how the objects were searched.   Recently TDirectoryFile was enhanced
by replacing a search through the whole list by a search through the bucket
resulting (without this patch) in a change of the ordering.