Add include guards for some glibc headers to make PCH work; capture a few more.
Headers can only be "read" from the PCH if they have a complete include guard.
Several glibc headers do not; the PCH cannot guarantee that in the context where
they are included at runtime, the parsing would yield the same result as at the
time the PCH was generated (with an include guard, the header would simply be
skipped). Thus, clang will parse the header; if that is different from the one
parsed at PCH generation time, the same header now has inconsistent size / file
locations etc - clang goes belly up. To solve this, we surround the original
glibc header with an explicit include guard surrounding an include_next (which
will then include the original glibc header).

wchar.h cannot be surrounded by include guards as it needs to be included
several times, with different preprocessor contexts. Instead capture it and
embed it in ROOT's runtime environment. For now, this only supports glibc in
/usr/include. Let's wait for the first request...