ROOT-7441 bug fix for call to non-static function
With this code:

struct X: public TObject {
  struct Y {
    ClassDef(Y,0)
  };
};

the interpreter version of ClassDef has a call to Error. Due to C++ lookup rules, the Error() found will be that of the enclosing X::Error() i.e. TObject::Error(), which cannot be called on an object of Y.
Instead, in the interpreted ClassDef version, call ::Error() explicitly - which is the right thing to do anyway, given that ClassDef() can be in a class that does not inherit from TObject.