Fix file corruption in TTeXDump::DrawPolyMarker
The current implementation of TTeXDump uses TVirtualPS::PrintFast based
methods to output TeX markup with automatic linewraps. Yet these methods
are optimized for PostScript format where there are a lot of space
characters that are used for newline placement. Current
TTeXDump::DrawPolyMarker would often produce a long contiguous lines
that trigger a forceful linewrap that can happen in the middle of real
number constant (ignored by latex) or even in the middle of latex
command (producing incorrect file).

One solution would be to rewrite TTeXDump using only PrintRaw (that you
can't mix with PrintStr/PrintFast/WriteReal). The other would be to fix
PrintFast to not introduce forced newline. The third option is less
intrusive and just adds additional spaces to provide clues for the
proper line wrapping (this is the one implemented in this change).