Fix ROOT-7500: crash due to change in base class which versioned derived class.
In the case where a base class does not have a version number and the
derived class does have a version number *and* that the schema of the base
class has changed but the version number of the derived class is not changed
then TStreamerInfo::CompareContent could not find the StreamerInfo of the base
class as TFile::ReadStreamerInfo had not beenloaded it yet.

In ROOT-7500, this situation inadvertently happens due to the use of the
typedef uint64_t which points to a different type on Mac and Linux (respectively
a 'unsigned long long' and a 'unsigned long').

See roottest/root/io/evolution/versions/execROOT7500.cxx