Integrate the commandline utils for root file inspection and modification
the idea behind these tools is to be able to perform on rootfiles operations
performed on a posix file system via the well known unix utilities.
An exception is roobrowse and rooprint, which allow to open a file and a
TBrowser and to plot all objects in a file respectively.