Start testing notebooks
The notebook is run locally and the output is saved as a notebook. Then the
output notebook is compared with the original one using the python difflib
module. If the notebooks do not differ modulo

* the welcome banner
* the new line at the end of the file

the test passes.

Two tiny notebooks have been added to the directory. They are small, about
1.5 kB big. This is size is comparable with simple C++ test programs but
notebooks can get quickly quite big if one takes into account markdown cells,
inlined images, jsROOT visualisation. For this reason it would be desirable
to download with cmake platform independent commands the notebooks from the web.
The list of notebooks is written down explicitely with the goal of avoiding
tests silently "disappearing" and not failing in presence of download issues.