Build the libHist dictionary with modules on.
We need to forward declare the Streamer function, because while
building the module it instantiates all pending instantiations.
This doesn't allow the dictionary to define the streamer function,
yielding a definition of a specialization after the implicit
instantiation.

With modules turned off, we managed to get away with it probably
because the instantiation was delayed until the end of TU or they
didn't end up in the same TU.