RooFit unit tests/tutorials: avoid pathological behaviour
last patch to RooAbsReal exposes pathological behaviour in the unit
tests (beta pdf at x = 0.0 can evaluate to zero probability, depending
on value of parameters, the "delayed" error reporting to avoid
evaluation error list overflow in RooAbsReal exposes this). The fix is
to change the lower bound of the variable x from 0.0 to 1e-9.