First snapshot of ROOT 7 interface personality.
We are not done for years to come - but this is a start.

It shows what interfaces (and parts of its implementations) can look like for
histograms, drawing, name registration etc. Many parts are still missing, notably
titles, histograms crunching away on lambdas, and hiding some of the TCoopPtr-s
inside these classes.

The goals for this new interface style are:

- Speed-up!
  - remove virtual function calls
  - reduce run-time complexity at the cost of compile-time, using inlining and type properties instead of object properties

- Robustness!
  - type-safety
  - ownership through types (smart pointers, references etc), reducing memory issues
  - compiler checking of options

- Simplicity!
  - smaller, focused interfaces: separate key/value registration, drawing etc
  - distinct non-public classes
  - use standard types instead of custom re-implementations: std::vector, double etc

- Interoperability!
  - play nice with current coding guidelines, for instance using standard library concepts
  - move types into ROOT namespace; headers into ROOT/ folder; libraries libROOTxyz

The build system integration will come during the coming days.