Fix crash due to API misuse ‘uncovered’ by change in default palette.
Probably due to the unlatching naming of functions, the code was using gROOT->GetColor when it meant
to use Color::GetPaletteColor.

gStyle->GetNumberOfColors() : returns the number of color in the current palette.
gROOT->GetColor(int) : returns the color with the requested index in the global list of color (non-contiguous)
TColor::GetNumberOfColors() : returns the number of color in the current palette.
TColor::GetColorPalette(int): return the color from the request slot in the current palette (contiguous 0 through TColor::GetNumberOfColors())

Per decision made in TColor::CreateColorWheel (gather the color number around fixed) large index (400, 600)
and TColor::InitializeColors (skip indices between 100 and 200), the list of color indices is not contiguous.

Previously the code worked because the default’s palette color also happened to be the first registered colors.