Set checksum for collection's StreamerInfo
Previously the checksum of the StreamerInfo of any collection (a
class which has a CollectionProxy) was set to zero.  However in the
case where the class was a custom collection (rather than an STL
collection) with only a CollectionProxy (and no custom streamer)
the class's checksum was stored/used during the streaming and
thus at read time (if the checksum changed) the StreamerInfo
could not be found.

Unlike STL collection, custom collection ought to indeed have
a version/checksum information stored as they could have older
version/implementation that did not rely on the CollectionProxy
for storage.