Resolve memory leak by leveraging move semantic. Fix ROOT-7683.
TSchemaRuleSet was return a new collection to be deleted by the caller and, of course, a few callers were missing
the delete.  Rather than explicitly adding the delete statement leverage the fact that we can cheaply return a
std::vector by value.   This has also the advantage of marking the content as const as intended.