allow Apple's 10.11 python to load again all ROOT shared libraries.
Due to the new way Apple builds python on 10.11 (El Capitan) each plugin
must specify where to find dependent libraries. This is done by
specifying the rpath using @loader_path (see man dyld) when creating
the plugins. In the past we provided this rpath specification only for
the executables but now for all plugins.