Improve the TMVA::Timer
The TMVA::Timer::DrawProgressBar( Int_t icounts, const TString& comment) function
used to be very expensive, it would redraw the bar every time it was called.

Now the previous state of the progress bar is cached so the the progress bar is
only redrawn when the output will look different.

Now that the timer is fast, add it to the TestRegression method.