Enable thread safety and awareness through ROOT::EnableMT()
The threading functionality remains encapsulated in libThread but
the thread safety and awareness are enabled via a method in Core:
no library dependency is introduced.
The usage of TThread::Initialize and Root.UseThreads have been
brovided of ROOT::EnableMT() and Root.EnableMT.
The usages of TThread::Initialize within ROOT have been adapted
to the new convention.