CMake: remove CMAKE_TINFO_LIBS
There is no direct usage of terminfo library (ncurses) in ROOT. LLVM
uses it for color support. If LLVM detects tinfo library, tinfo will be
added for rootcling_tmp, rootcling, and libCling.so targets. Thus
explicit check for tinfo is not required.

This allows to use -DLLVM_ENABLE_TERMINFO=OFF to disable ncurses for
ROOT.

Signed-off-by: David Abdurachmanov <David.Abdurachmanov@cern.ch>