Thread: revised implementation for TSemaphore
The existing implementation of TSemaphore had a problem due to the fact that no track was kept of
the number of waiting threads. If subsequent posts occured fast enough they were or could be lost.
This patch revises the logic, adding a counter for waiting threads. The implementation is based on
STL primitives, following the new policy of the team.