Add modules support for opengl.
One wrinkle is glew.h. It needs to #include glu.h (which internally #includes
gl.h) without including gl.h. This is done by defining gl's include guards in
glew.h. In order to make this case work we have to use config_macros. This
doesn't work because of a bug/feature in the modules. The problems are triggered
on osx because OpenGL is shipped with its own module map glew has to work with
the OpenGL top module.

We can workaround the issue by leveraging the buildsystem. Before starting to
build the dependents on glew.h libraries, we need to prebuild a module for
glew with OpenGL modulemap disabled.

The trick doesn't work on cmake yet: one of the reasons is that it seems to
build the dependents on glew with various different compiler flags. In this case
the prebuilt glew module's flags differ from the ones of the its dependents.
to CMake doesn't work