Fix error due to concurrent access to history sqlite db
nbconvert apparently does not seem to transmit to the ipython instance
which runs the notebook the parameters of the configuration passed
to it, especially the history file.
The workaround is to create a temp directory per notebook where the
profile will be written and to communicate the name of this directory
to ipython via the IPYTHONDIR environment variable.