Fix ROOT-7864: c++ exception in a constructor is not propagated to python
This is due to a change of the exception mapping code from Toby (GSoC student).
In TMethodHandler::Execute, exceptions are turned into magic instead of null for
all cases now, and TConstructorHolder (nor the other XYZHolders) take this into
account. (Only TMethodHolder does.)

Following the white rabbit I noticed that the combination of both TPy and TPyCPP
magics should always occur together, but at least on one path doesn't and on
another has a dead branch that needs pruning. If the code keeps these paths, at
least both should be captured in a single helper IsMagic() to prevent such
mishaps. (Some code is there in ObjectProxy but not widely used.)

And anyway, it doesn't actually matter, as the magic is not treated differently
from NULL up from TMethodHolder (even as the opposite is true), so might as well
strip all of that out and use a null-check together with PyErr_Occurred() to
implement the code paths (and since the latter is already there in all relevant
cases, the magic really is superfluous). This especially since the original code
was already broken in case the ctor called python code that threw. (There would
be a small performance benefit over checking the magic v.s. PyErr_Occurred(),
but that'd be all.)