Do not process a cell if curly braces are not balanced
at the root prompt, one can type code on multiple lines, for example in presence
of open scopes.
In a notebook this behaviour is not well defined: each cell is somehow self
contained. This commit enforces all cells to be self contained, checking the
error code of TInterpreter::ProcessLine, which is TInterpreter::kProcessing
when the closing curly braces do not equal the opened ones.