Fix ROOT-7901 by making all static init function name *really* unique.
See also ROOT-7775.
Extends/finishes d84ee9cc74fde98f878148d16f253dec99263022

The comments in the patch to fix ROOT-7775 pointed out that the problem
was that the names of the __cxx_global_var_init* functions were not unique
across modules. As a fix, he added the module name, so the function name
is formed from "__cxx_global_var_init" + module-name + serial-number.
However — this is not sufficient to make the names unique, because
the module-name ends with digits. So for example, serial number
29 in cling-module-45 would have the same name
as serial number 9 in cling-module-452!