Fix unnecessary autoparse on an abstract class.
Remove the call to HasDefaultConstructor in the case that the
class is loaded and does not have an ioctor wrapper.  The call
causes an autoparse and the dictionary generator has already
made this check and would have created the ioctor wrapper if
there was a default or i/o constructor to be found.  The most
common cause of this is a dictionary for an abstract base
class (which cannot be constructed).

Philippe Canal reviewed this patch.