Provice std::hash for TMVA::Event* and TMVA::DecisionTreeNode*
It seems the default std::hash functor for these types is not optimal.
This patch suggests to right shift the pointer value before passing
it on to std::hash<size_t>. This improves speed on my system because
the rightmost bits of the pointers are always zero and somehow the
default hash implementation doesn't like that.

There might be better ideas to improve the maps of these pointers...

Warning: This is compiler/architecture dependent!
Tested on 64 bit linux using gcc 4.8.4