Add support for TTreeCloner cache in TTree::CopyEntries.
The cache is turned on by default.  To change it size add to the option string:
   cachesize=SIZE
where SIZE is in 'number of bytes' and can be expressed in 'human readable form' (number followed
by size unit like MB, MiB, GB or GiB, etc..
or set SIZE to zero to disable it.