Add new mode in CanSplit. Avoid spurrious auto-parse in TClass::CanSplit. This fixes ROOT-7972.
Add a new mode for fCanSplit (2) which indicates that this class and any derived class
should not be split.

Rework the mechanism checking the base classes.  Instead of using InheritsFrom, which
lead in some cases, including the case where the class derived from an STL collection,
to spurrious autoparsing (to look at the base class of the collection!), we use a custom
walk through the tree of base classes that checks their value of fCanSplit.

This also has the side-effect of allowing the extension of the concept 'base class that
prevent its derived class from being split' to any user class.