Implement IO for unique_ptr.
this commit implements the functionality necessary to perform I/O operations,
both column- and row-wise with std::unique_ptr<T> instances as if they were
simple, good old pointers.
Only unique_ptrs with a default second argument can be treated.
The implementation assumes that the layout in memory of a unique_ptr<T> is the
same of the one of a T*.

Some details:
- TStreamer infor construction, both starting from the type system and from files
(i.e. TStreamerInfo Build and BuildOld methods) have been adapted to accommodate
the I/O of unique_ptrs.
- Dictionary generation stops if a persistent data member is a unique_ptr with a
non default delete argument.