Fix handling of deprecated options
The configure script currently uses a simple grep to determine if an option
is deprecated. This causes problems, since the enable-reflex option is
deprecated and grepping for enable-r will result in a match. The configure
script therefore considers enable-r deprecated too, even though it is not.
This commit addresses this issue.