Don't use hexfloat constant
Hexfloat constants are part of c++17. They can not be used when
compileing with -std=c++11 or -std=c++14. Doing so will result in
"error: exponent has no digits".

It would be possible to use them with GNU extensions enabled (i.e.
with -std=gnu++11 or -std=gnu++14), but root compilation is not
configured this way.