Reduce the needed memory for compilation
The linking of rootcling and libCling requires a lot of memory. Since these
are linke form mostly the same objects, the build is ready to link them at
the same time. If you make a parallel build this means that the two targets
that require the most amount of memory are being linked in parallel. This
exhausts the available memory, and the computer starts swapping.

This adds a dependency of one of the targets to the other. The dependency is
not really there since it is not needed for building, but it prevents the
two memory consuming targets to be built in parallel.