Remove some Begin_Html/End_Html tags
This is not an attempt to convert the documentation to doxygen format.
It is rather an attempt to address leftovers from a partial (automated?)
convertion that left some of the Begin_Html/End_Html tags behind. Tags
that now show up verbatim in the generated doxygen documentation.

Most of these files still requires proper doxygen conversion.