Avoid leaking a TTreeFormulaManager when there is no formula.
See https://root.cern.ch/phpBB3/viewtopic.php?t=21462.

The owner of the TTreeFormulaManager is actually the TTreeFormula themselves and when there is no
formula there was nobody to delete the manager.  This is an obvious place where shared_ptr would be
making the code better.  However in this case, the creation of the TTreeFormulaManager was spurrious and
redundant anyway, so this is the straightforward fix.