Fix module builds of includers of glew.h and TGLIncludes.
glew.h includes glu.h. glu.h includes gl.h. However glew.h defines the inclusion
protector of gl.h disallowing gl.h in the same TU. By design, glew.h is meant to
replace gl.h.

Modules don't allow that, because OpenGL comes with modules support on MacOS.
Whenever glew.h tries to do this trick on a module-aware OpenGL framework it
fails.

With some great help by the clang modules community there is a "less bad" way to
fix this issue. Whenever we build the system module for OpenGL we swap gl.h with
our own (empty) version so that it is not present in the pcm. This way glew.h
can do its job. Another advantage of the approach is that we can diagnose when
a ROOT user uses gl.h directly and warn him to use the specially crafted for
that purpose TGLIncludes.h